app/requirements.txt
flask
flask-sqlalchemy
Flask-Login
Flask-JWT-Extended
flask-swagger-ui
marshmallow
pytest
pytest-flask
flask_cors
pytz
